+++ EstatisticaFatec.Tests/VariavelContinuaTest.cs
@@ -10,7 +10,7 @@ namespace EstatisticaFatec.Tests
         [TestMethod]
         public void Input()
         {
-            var numeros = new List<int>
+            var numeros = new List<decimal>
             {
                 20,37,40,32,28,26,25,32,39,40,50,47,46,31,24,21,35,39,38,43
             };
